requirements.txt
Django==5.2.7
channels==4.1.0
channels-redis==4.3.0
daphne==4.1.2
django-crispy-forms==2.3
crispy-tailwind==0.5.0
pillow==10.4.0
pycryptodome==3.21.0
redis==5.2.0
celery==5.4.0